==> 0.Diagnosis_ICD10.py <==
# ICD10
# 按 ['F32', 'F33']  疾病代码定义MDD    对应MergerAllData.py中 Dataset 2
from collections import Counter
import time
import threading
from concurrent.futures import ThreadPoolExecutor, as_completed
from SUtils import *

# ICD10 Disease Columns - -41270
cols_ICD10 = ['41270-0.0', '41270-0.1', '41270-0.2', '41270-0.3', '41270-0.4', '41270-0.5', '41270-0.6', '41270-0.7', '41270-0.8', '41270-0.9', '41270-0.10', '41270-0.11', '41270-0.12', '41270-0.13', '41270-0.14', '41270-0.15', '41270-0.16', '41270-0.17', '41270-0.18', '41270-0.19', '41270-0.20', '41270-0.21', '41270-0.22', '41270-0.23', '41270-0.24', '41270-0.25', '41270-0.26', '41270-0.27', '41270-0.28', '41270-0.29', '41270-0.30', '41270-0.31', '41270-0.32', '41270-0.33', '41270-0.34', '41270-0.35', '41270-0.36', '41270-0.37', '41270-0.38', '41270-0.39', '41270-0.40', '41270-0.41', '41270-0.42', '41270-0.43', '41270-0.44', '41270-0.45', '41270-0.46', '41270-0.47', '41270-0.48', '41270-0.49', '41270-0.50', '41270-0.51', '41270-0.52', '41270-0.53', '41270-0.54', '41270-0.55', '41270-0.56', '41270-0.57', '41270-0.58', '41270-0.59', '41270-0.60', '41270-0.61', '41270-0.62', '41270-0.63', '41270-0.64', '41270-0.65', '41270-0.66', '41270-0.67', '41270-0.68', '41270-0.69', '41270-0.70', '41270-0.71', '41270-0.72', '41270-0.73', '41270-0.74', '41270-0.75', '41270-0.76', '41270-0.77', '41270-0.78', '41270-0.79', '41270-0.80', '41270-0.81', '41270-0.82', '41270-0.83', '41270-0.84', '41270-0.85', '41270-0.86', '41270-0.87', '41270-0.88', '41270-0.89', '41270-0.90', '41270-0.91', '41270-0.92', '41270-0.93', '41270-0.94', '41270-0.95', '41270-0.96', '41270-0.97', '41270-0.98', '41270-0.99', '41270-0.100', '41270-0.101', '41270-0.102', '41270-0.103', '41270-0.104', '41270-0.105', '41270-0.106', '41270-0.107', '41270-0.108', '41270-0.109', '41270-0.110', '41270-0.111', '41270-0.112', '41270-0.113', '41270-0.114', '41270-0.115', '41270-0.116', '41270-0.117', '41270-0.118', '41270-0.119', '41270-0.120', '41270-0.121', '41270-0.122', '41270-0.123', '41270-0.124', '41270-0.125', '41270-0.126', '41270-0.127', '41270-0.128', '41270-0.129', '41270-0.130', '41270-0.131', '41270-0.132', '41270-0.133', '41270-0.134', '41270-0.135', '41270-0.136', '41270-0.137', '41270-0.138', '41270-0.139', '41270-0.140', '41270-0.141', '41270-0.142', '41270-0.143', '41270-0.144', '41270-0.145', '41270-0.146', '41270-0.147', '41270-0.148', '41270-0.149', '41270-0.150', '41270-0.151', '41270-0.152', '41270-0.153', '41270-0.154', '41270-0.155', '41270-0.156', '41270-0.157', '41270-0.158', '41270-0.159', '41270-0.160', '41270-0.161', '41270-0.162', '41270-0.163', '41270-0.164', '41270-0.165', '41270-0.166', '41270-0.167', '41270-0.168', '41270-0.169', '41270-0.170', '41270-0.171', '41270-0.172', '41270-0.173', '41270-0.174', '41270-0.175', '41270-0.176', '41270-0.177', '41270-0.178', '41270-0.179', '41270-0.180', '41270-0.181', '41270-0.182', '41270-0.183', '41270-0.184', '41270-0.185', '41270-0.186', '41270-0.187', '41270-0.188', '41270-0.189', '41270-0.190', '41270-0.191', '41270-0.192', '41270-0.193', '41270-0.194', '41270-0.195', '41270-0.196', '41270-0.197', '41270-0.198', '41270-0.199', '41270-0.200', '41270-0.201', '41270-0.202', '41270-0.203', '41270-0.204', '41270-0.205', '41270-0.206', '41270-0.207', '41270-0.208', '41270-0.209', '41270-0.210', '41270-0.211', '41270-0.212', '41270-0.213', '41270-0.214', '41270-0.215', '41270-0.216', '41270-0.217', '41270-0.218', '41270-0.219', '41270-0.220', '41270-0.221', '41270-0.222', '41270-0.223', '41270-0.224', '41270-0.225', '41270-0.226', '41270-0.227', '41270-0.228', '41270-0.229', '41270-0.230', '41270-0.231', '41270-0.232', '41270-0.233', '41270-0.234', '41270-0.235', '41270-0.236', '41270-0.237', '41270-0.238', '41270-0.239', '41270-0.240', '41270-0.241', '41270-0.242']

# ICD10-Affective-Disorder  Disease Code
code_ICD10_MDD = ['F32', 'F33']  # MDD 1


# 判断字符串result是否以某个子串tag开头
def startCmp(listTag, listResult):
    for result in listResult:
        for tag in listTag:
            if result.startswith(tag):
                return 1
    return 0


def singe_row(x):
    result_ICD10 = [str(elem) if not pd.isna(elem) else None for elem in x[cols_ICD10]]
    result_ICD10 = list(filter(None, result_ICD10))
    return startCmp(code_ICD10_MDD, result_ICD10)


def get_AD(i):
    df = pd.read_csv(data_path +"ICD/" + str(i) + ".csv", low_memory=False)
    df['ICD'] = df.apply(lambda row: singe_row(row), axis=1)
    if not os.path.exists(data_path +"ICD_Cate/"):
        os.makedirs(data_path +"ICD_Cate/")
    df.to_csv(data_path +"ICD_Cate/" + str(i) + ".csv", index=None)
    return Counter(df['ICD'].tolist())  # dict


def get_AD_multiprocess():
    MDD = 0
    start_time = time.time()
    with ThreadPoolExecutor(6) as executor:  # 创建 ThreadPoolExecutor
        future_list = [executor.submit(get_AD, i) for i in range(17)]  # 提交任务

    for future in as_completed(future_list):
        result = future.result()  # 获取任务结果
        print("%s get result : %s" % (threading.current_thread().getName(), result))
        MDD += result[1]
    print("Affective Disorder Number: 1:{}".format(MDD))
    print('%s cost %d second' % (threading.current_thread().getName(), time.time() - start_time))


def split_data():
    df_chunk = pd.read_csv(root_path+"Cate_2002.csv", chunksize=30000, low_memory=False, usecols=['eid']+cols_ICD10)

    if not os.path.exists(data_path +"ICD/"):
        os.makedirs(data_path +"ICD/")
    for i, chunk in enumerate(df_chunk):
        print("get Data by chunk {}: size {}".format(i, chunk.shape[0]))
        chunk.to_csv(data_path +"ICD/" + str(i) + ".csv", index=None)

def merge_data():
    df_list = []
    for i in range(17):
        df = pd.read_csv(data_path +"ICD_Cate/"+str(i)+".csv", usecols=['eid', 'ICD'])
        df_list.append(df)
    df_all = pd.concat(df_list)
    df_all.to_csv(data_path + "ICD_Cate/" + "ICD.csv", index=None)

if __name__ == '__main__':
    split_data()
    get_AD_multiprocess()
    merge_data()
