==> 0.Medicine_filter.py <==
# 统计 Antidepressant 和 antipsychotics， 用于排除Normal Control
# Antidepressant Usage(Glanville et al., 2021a)
# The criteria for defining ‘Antidepressant Usage’ cases:
# •	Self-reported taking antidepressant medication at baseline or the subsequent two repeat assessments [ID 20003]. Antidepressant codes [UKB Data-Coding 4]: 1140879616, 1140921600, 1140879540, 1140867878, 1140916282, 1140909806, 1140867888, 1141152732, 1141180212, 1140879634, 1140867876, 1140882236, 1141190158, 1141200564, 1140867726,  1140879620, 1140867818, 1140879630, 1140879628, 1141151946, 1140867948, 1140867624, 1140867756, 1140867884, 1141151978, 1141152736, 1141201834, 1140867690, 1140867640, 1140867920, 1140867850, 1140879544, 1141200570, 1140867934, 1140867758, 1140867914, 1140867820, 1141151982, 1140882244, 1140879556, 1140867852, 1140867860, 1140917460, 1140867938, 1140867856, 1140867922, 1140910820, 1140882312, 1140867944, 1140867784, 1140867812, 1140867668
# additionally including (Data Field 20003)
# 1140879624, 1140879632, 1140879552, 1140916288, 1140867916, 1140867806, 1140867824, 1140867712, 1140867720, 1140867722, 1140867734

# For antipsychotics, we identified codes similar to Glanville et al. (Glanville et al., 2021a).
# The criteria for defining ‘Antipsychotic Usage’ cases:
# •	Self-reported taking antipsychotic medication at baseline or the subsequent two repeat assessments [ID 20003]. Antipsychotic codes [UKB Data-Coding 4]: 1140868170, 1140928916, 1141152848, 1140867444, 1140879658, 1140868120, 1141153490, 1140867304, 1141152860, 1140867168, 1141195974, 1140867244, 1140867152, 1140909800, 1140867420, 1140879746, 1141177762, 1140867456, 1140867952, 1140867150, 1141167976, 1140882100, 1140867342, 1140863416, 1141202024, 1140882098, 1140867184, 1140867092, 1140882320, 1140910358, 1140867208, 1140909802, 1140867134, 1140867306, 1140867210, 1140867398, 1140867078, 1140867218, 1141201792, 1141200458, 1140867136, 1140879750, 1140867180, 1140867546, 1140928260, 1140927956.
from SUtils import *


code_antidepressant = [1140879616, 1140921600, 1140879540, 1140867878, 1140916282, 1140909806, 1140867888, 1141152732, 1141180212, 1140879634, 1140867876, 1140882236, 1141190158, 1141200564, 1140867726, 1140879620, 1140867818, 1140879630, 1140879628, 1141151946, 1140867948, 1140867624, 1140867756, 1140867884, 1141151978, 1141152736, 1141201834, 1140867690, 1140867640, 1140867920, 1140867850, 1140879544, 1141200570, 1140867934, 1140867758, 1140867914, 1140867820, 1141151982, 1140882244, 1140879556, 1140867852, 1140867860, 1140917460, 1140867938, 1140867856, 1140867922, 1140910820, 1140882312, 1140867944, 1140867784, 1140867812, 1140867668, 1140879624, 1140879632, 1140879552, 1140916288, 1140867916, 1140867806, 1140867824, 1140867712, 1140867720, 1140867722, 1140867734]
code_antidepressant = [str(elem) for elem in code_antidepressant]
code_antipsychotics = [1140868170, 1140928916, 1141152848, 1140867444, 1140879658, 1140868120, 1141153490, 1140867304, 1141152860, 1140867168, 1141195974, 1140867244, 1140867152, 1140909800, 1140867420, 1140879746, 1141177762, 1140867456, 1140867952, 1140867150, 1141167976, 1140882100, 1140867342, 1140863416, 1141202024, 1140882098, 1140867184, 1140867092, 1140882320, 1140910358, 1140867208, 1140909802, 1140867134, 1140867306, 1140867210, 1140867398, 1140867078, 1140867218, 1141201792, 1141200458, 1140867136, 1140879750, 1140867180, 1140867546, 1140928260, 1140927956]
code_antipsychotics = [str(elem) for elem in code_antipsychotics]

cols_names = ['20003-0.0', '20003-0.1', '20003-0.2', '20003-0.3', '20003-0.4', '20003-0.5', '20003-0.6', '20003-0.7', '20003-0.8', '20003-0.9', '20003-0.10', '20003-0.11', '20003-0.12', '20003-0.13', '20003-0.14', '20003-0.15', '20003-0.16', '20003-0.17', '20003-0.18', '20003-0.19', '20003-0.20', '20003-0.21', '20003-0.22', '20003-0.23', '20003-0.24', '20003-0.25', '20003-0.26', '20003-0.27', '20003-0.28', '20003-0.29', '20003-0.30', '20003-0.31', '20003-0.32', '20003-0.33', '20003-0.34', '20003-0.35', '20003-0.36', '20003-0.37', '20003-0.38', '20003-0.39', '20003-0.40', '20003-0.41', '20003-0.42', '20003-0.43', '20003-0.44', '20003-0.45', '20003-0.46', '20003-0.47', '20003-1.0', '20003-1.1', '20003-1.2', '20003-1.3', '20003-1.4', '20003-1.5', '20003-1.6', '20003-1.7', '20003-1.8', '20003-1.9', '20003-1.10', '20003-1.11', '20003-1.12', '20003-1.13', '20003-1.14', '20003-1.15', '20003-1.16', '20003-1.17', '20003-1.18', '20003-1.19', '20003-1.20', '20003-1.21', '20003-1.22', '20003-1.23', '20003-1.24', '20003-1.25', '20003-1.26', '20003-1.27', '20003-1.28', '20003-1.29', '20003-1.30', '20003-1.31', '20003-1.32', '20003-1.33', '20003-1.34', '20003-1.35', '20003-1.36', '20003-1.37', '20003-1.38', '20003-1.39', '20003-1.40', '20003-1.41', '20003-1.42', '20003-1.43', '20003-1.44', '20003-1.45', '20003-1.46', '20003-1.47', '20003-2.0', '20003-2.1', '20003-2.2', '20003-2.3', '20003-2.4', '20003-2.5', '20003-2.6', '20003-2.7', '20003-2.8', '20003-2.9', '20003-2.10', '20003-2.11', '20003-2.12', '20003-2.13', '20003-2.14', '20003-2.15', '20003-2.16', '20003-2.17', '20003-2.18', '20003-2.19', '20003-2.20', '20003-2.21', '20003-2.22', '20003-2.23', '20003-2.24', '20003-2.25', '20003-2.26', '20003-2.27', '20003-2.28', '20003-2.29', '20003-2.30', '20003-2.31', '20003-2.32', '20003-2.33', '20003-2.34', '20003-2.35', '20003-2.36', '20003-2.37', '20003-2.38', '20003-2.39', '20003-2.40', '20003-2.41', '20003-2.42', '20003-2.43', '20003-2.44', '20003-2.45', '20003-2.46', '20003-2.47', '20003-3.0', '20003-3.1', '20003-3.2', '20003-3.3', '20003-3.4', '20003-3.5', '20003-3.6', '20003-3.7', '20003-3.8', '20003-3.9', '20003-3.10', '20003-3.11', '20003-3.12', '20003-3.13', '20003-3.14', '20003-3.15', '20003-3.16', '20003-3.17', '20003-3.18', '20003-3.19', '20003-3.20', '20003-3.21', '20003-3.22', '20003-3.23', '20003-3.24', '20003-3.25', '20003-3.26', '20003-3.27', '20003-3.28', '20003-3.29', '20003-3.30', '20003-3.31', '20003-3.32', '20003-3.33', '20003-3.34', '20003-3.35', '20003-3.36', '20003-3.37', '20003-3.38', '20003-3.39', '20003-3.40', '20003-3.41', '20003-3.42', '20003-3.43', '20003-3.44', '20003-3.45', '20003-3.46', '20003-3.47']
def single_row(x):
    result_list = [str(int(elem)) if not np.isnan(elem) else None for elem in x[cols_names]]
    result_list = list(filter(None, result_list))  # 去除列表空值
    antidepressant_code = list(set(code_antidepressant) & set(result_list)) # 求交集
    antipsychotics_code = list(set(code_antipsychotics) & set(result_list))
    antidepressant = 1 if len(antidepressant_code) > 0 else 0
    antipsychotics = 1 if len(antipsychotics_code) > 0 else 0
    print("result_list: {}, antidepressant: {}, antidepressant_code: {}, antipsychotics: {}, antipsychotics_code: {}".format(result_list, antidepressant, antidepressant_code, antipsychotics, antipsychotics_code))
    return antidepressant, antidepressant_code, antipsychotics, antipsychotics_code


def m_intake():
    # df = pd.read_csv(root_path + "Cate_3001.csv")
    df = pd.read_csv(root_path + "Cate_3001.csv", usecols=['eid']+cols_names)
    # df_columns = df.columns.tolist()
    # print(df.columns.tolist())
    row = df.apply(lambda x: single_row(x), axis=1)
    antidepressant, antidepressant_code, antipsychotics, antipsychotics_code = zip(*row)
    df['antidepressant'] = pd.Series(antidepressant)
    df['antidepressant_code'] = pd.Series(antidepressant_code)
    df['antipsychotics'] = pd.Series(antipsychotics)
    df['antipsychotics_code'] = pd.Series(antipsychotics_code)
    if not os.path.exists(data_path +"Medicine_intake/"):
        os.makedirs(data_path +"Medicine_intake/")
    df.to_csv(data_path + "Medicine_intake/" +"Medicine_intake.csv", columns=['eid', 'antidepressant', 'antidepressant_code', 'antipsychotics', 'antipsychotics_code'], index=None)

if __name__ == '__main__':
    m_intake()
